+++ packages/core/select/src/AsyncSelect.js
@@ -1,6 +1,0 @@
-// @flow
-
-import AsyncSelect from 'react-select/async';
-import createSelect from './createSelect';
-
-export default createSelect(AsyncSelect);