+++ packages/elements/share/src/components/ShareDialogContainer.tsx
@@ -484,7 +484,7 @@ export class ShareDialogContainerInternal extends React.Component<
                 <FormattedMessage {...messages.inviteTriggerButtonText} />
               )
             }
-            product={product}
+            product={product!}
           />
         </MessagesIntlProvider>
       </ErrorBoundary>