+++ packages/navigation/atlassian-switcher/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/atlassian-switcher
 
+## 4.6.9
+
+### Patch Changes
+
+- [patch][ea75c17b3a](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ea75c17b3a):
+
+  internal typescript fixes
+
 ## 4.6.8
 
 ### Patch Changes